clean up response db
