+++ src/main/java/teammates/storage/api/FeedbackResponsesDb.java
@@ -22,7 +22,6 @@ import teammates.common.exception.EntityDoesNotExistException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Assumption;
 import teammates.common.util.Const;
-import teammates.common.util.Logger;
 import teammates.storage.entity.FeedbackResponse;
 
 /**
@@ -33,67 +32,36 @@ import teammates.storage.entity.FeedbackResponse;
  */
 public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackResponseAttributes> {
 
-    private static final Logger log = Logger.getLogger();
-
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return Null if not found.
+     * Gets a feedback response.
      */
     public FeedbackResponseAttributes getFeedbackResponse(String feedbackResponseId) {
-        return makeAttributesOrNull(getFeedbackResponseEntityWithCheck(feedbackResponseId));
-    }
-
-    /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return Null if not found.
-     */
-    public FeedbackResponseAttributes getFeedbackResponse(
-            String feedbackQuestionId, String giverEmail, String receiverEmail) {
-        return makeAttributesOrNull(getFeedbackResponseEntityWithCheck(feedbackQuestionId, giverEmail, receiverEmail));
-    }
-
-    /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return Null if not found.
-     */
-    private FeedbackResponse getFeedbackResponseEntityWithCheck(String feedbackResponseId) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, feedbackResponseId);
 
         FeedbackResponse fr = getFeedbackResponseEntity(feedbackResponseId);
-        if (fr == null) {
-            log.info("Trying to get non-existent response: " + feedbackResponseId + ".");
-            return null;
-        }
-        return fr;
+
+        return makeAttributesOrNull(fr);
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return Null if not found.
+     * Gets a feedback response by unique constraint question-giver-receiver.
      */
-    private FeedbackResponse getFeedbackResponseEntityWithCheck(
+    public FeedbackResponseAttributes getFeedbackResponse(
             String feedbackQuestionId, String giverEmail, String receiverEmail) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, feedbackQuestionId);
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, giverEmail);
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, receiverEmail);
 
-        FeedbackResponse fr = getFeedbackResponseEntity(feedbackQuestionId, giverEmail, receiverEmail);
-        if (fr == null) {
-            log.warning("Trying to get non-existent response: " + feedbackQuestionId + "/" + "from: " + giverEmail
-                    + " to: " + receiverEmail);
-            return null;
-        }
-        return fr;
+        FeedbackResponse fr =
+                getFeedbackResponseEntity(FeedbackResponse.generateId(feedbackQuestionId, giverEmail, receiverEmail));
+
+        return makeAttributesOrNull(fr);
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all feedback responses of a question in a specific section.
+     *
+     * <p>{@code sectionDetail} specifies the criteria of classifying a response in a section</p>
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForQuestionInSection(
             String feedbackQuestionId, String section, SectionDetail sectionDetail) {
@@ -105,33 +73,29 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all feedback responses for a question.
      */
-    public List<FeedbackResponseAttributes> getFeedbackResponsesForQuestion(
-            String feedbackQuestionId) {
+    public List<FeedbackResponseAttributes> getFeedbackResponsesForQuestion(String feedbackQuestionId) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, feedbackQuestionId);
 
         return makeAttributes(getFeedbackResponseEntitiesForQuestion(feedbackQuestionId));
     }
 
     /**
-     * Finds the responses for a specified question within a given range.
-     *
-     * @return An empty list if no such responses are found.
+     * Checks whether there are responses for a question.
      */
-    public List<FeedbackResponseAttributes> getFeedbackResponsesForQuestionWithinRange(
-            String feedbackQuestionId, int range) {
+    public boolean areThereResponsesForQuestion(String feedbackQuestionId) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, feedbackQuestionId);
 
-        return makeAttributes(getFeedbackResponseEntitiesForQuestionWithinRange(feedbackQuestionId, range));
+        return !load()
+                .filter("feedbackQuestionId =", feedbackQuestionId)
+                .limit(1)
+                .list()
+                .isEmpty();
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses of a feedback session in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSession(
             String feedbackSessionName, String courseId) {
@@ -142,9 +106,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses of a feedback session in a course with limit.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionWithinRange(
             String feedbackSessionName, String courseId, int range) {
@@ -155,9 +117,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to/from a section in a feedback session in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionInSection(
             String feedbackSessionName, String courseId, String section) {
@@ -168,6 +128,9 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
         return makeAttributes(getFeedbackResponseEntitiesForSessionInSection(feedbackSessionName, courseId, section));
     }
 
+    /**
+     * Gets all responses given to&from a section in a feedback session in a course.
+     */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionInGiverAndRecipientSection(
             String feedbackSessionName, String courseId, String section) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, feedbackSessionName);
@@ -179,9 +142,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given from a section in a feedback session in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionFromSection(
             String feedbackSessionName, String courseId, String section) {
@@ -193,9 +154,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to a section in a feedback session in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionToSection(
             String feedbackSessionName, String courseId, String section) {
@@ -207,9 +166,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to/from a section in a feedback session in a course with limit.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionInSectionWithinRange(
             String feedbackSessionName, String courseId, String section, int range) {
@@ -223,9 +180,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given from a section in a feedback session in a course with limit.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionFromSectionWithinRange(
             String feedbackSessionName, String courseId, String section, int range) {
@@ -239,9 +194,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to a section in a feedback session in a course with limit.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForSessionToSectionWithinRange(
             String feedbackSessionName, String courseId, String section, int range) {
@@ -255,9 +208,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to a user for a question.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForReceiverForQuestion(
             String feedbackQuestionId, String receiver) {
@@ -268,9 +219,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to a user and given by a section for a question.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForReceiverForQuestionInSection(
             String feedbackQuestionId, String receiver, String section) {
@@ -283,9 +232,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given by a user for a question.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesFromGiverForQuestion(
             String feedbackQuestionId, String giverEmail) {
@@ -296,9 +243,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given by a user and given to a section for a question.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesFromGiverForQuestionInSection(
             String feedbackQuestionId, String giverEmail, String section) {
@@ -311,9 +256,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     *  Preconditions: <br>
-     * * All parameters are non-null.
-     *  @return An empty list if no such responses are found.
+     * Gets all responses given by a user in a feedback session of a course with limit.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesFromGiverForSessionWithinRange(
             String giverEmail, String feedbackSessionName, String courseId, int range) {
@@ -326,9 +269,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given to a user in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesForReceiverForCourse(
             String courseId, String receiver) {
@@ -339,9 +280,7 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     /**
-     * Preconditions: <br>
-     * * All parameters are non-null.
-     * @return An empty list if no such responses are found.
+     * Gets all responses given by a user in a course.
      */
     public List<FeedbackResponseAttributes> getFeedbackResponsesFromGiverForCourse(
             String courseId, String giverEmail) {
@@ -442,24 +381,11 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
      */
     public boolean hasFeedbackResponseEntitiesForCourse(String courseId) {
         Assumption.assertNotNull(Const.StatusCodes.DBLEVEL_NULL_INPUT, courseId);
-        return !getFeedbackResponseEntitiesForCourseWithinRange(courseId, 1).isEmpty();
-    }
-
-    private List<FeedbackResponse> getFeedbackResponseEntitiesForCourseWithinRange(String courseId, int range) {
-        return load().filter("courseId =", courseId).limit(range).list();
+        return !load().filter("courseId =", courseId).limit(1).list().isEmpty();
     }
 
     private FeedbackResponse getFeedbackResponseEntity(String feedbackResponseId) {
         return load().id(feedbackResponseId).now();
-    }
-
-    private FeedbackResponse getFeedbackResponseEntity(
-            String feedbackQuestionId, String giverEmail, String receiver) {
-        return load()
-                .filter("feedbackQuestionId =", feedbackQuestionId)
-                .filter("giverEmail =", giverEmail)
-                .filter("receiver =", receiver)
-                .first().now();
     }
 
     private Set<FeedbackResponse> getFeedbackResponseEntitiesForQuestionInSection(
@@ -505,17 +431,12 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
     }
 
     private List<FeedbackResponse> getFeedbackResponseEntitiesForQuestion(String feedbackQuestionId) {
-        return getFeedbackResponseEntitiesForQuestionWithinRange(feedbackQuestionId, -1);
-    }
-
-    private List<FeedbackResponse> getFeedbackResponseEntitiesForQuestionWithinRange(String feedbackQuestionId, int range) {
         return load()
                 .filter("feedbackQuestionId =", feedbackQuestionId)
-                .limit(range + 1).list();
+                .list();
     }
 
-    private List<FeedbackResponse> getFeedbackResponseEntitiesForSession(
-            String feedbackSessionName, String courseId) {
+    private List<FeedbackResponse> getFeedbackResponseEntitiesForSession(String feedbackSessionName, String courseId) {
         return getFeedbackResponseEntitiesForSessionWithinRange(feedbackSessionName, courseId, -1);
     }
 
@@ -546,8 +467,24 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
 
     private List<FeedbackResponse> getFeedbackResponseEntitiesForSessionInGiverAndRecipientSection(
             String feedbackSessionName, String courseId, String section) {
-        return getFeedbackResponseEntitiesForSessionInGiverAndRecipientSectionWithinRange(feedbackSessionName, courseId,
-                section, -1);
+        List<FeedbackResponse> feedbackResponses = new ArrayList<>();
+
+        feedbackResponses.addAll(load()
+                .filter("feedbackSessionName = ", feedbackSessionName)
+                .filter("courseId =", courseId)
+                .filter("giverSection =", section)
+                .filter("receiverSection =", section)
+                .list());
+
+        // also show responses in section with giver but without recipient
+        feedbackResponses.addAll(load()
+                .filter("feedbackSessionName = ", feedbackSessionName)
+                .filter("courseId =", courseId)
+                .filter("giverSection =", section)
+                .filter("receiverSection =", "None")
+                .list());
+
+        return feedbackResponses;
     }
 
     private List<FeedbackResponse> getFeedbackResponseEntitiesForSessionFromSection(
@@ -575,28 +512,6 @@ public class FeedbackResponsesDb extends EntitiesDb<FeedbackResponse, FeedbackRe
         }
 
         return feedbackResponses.values();
-    }
-
-    private List<FeedbackResponse> getFeedbackResponseEntitiesForSessionInGiverAndRecipientSectionWithinRange(
-            String feedbackSessionName, String courseId, String section, int range) {
-        List<FeedbackResponse> feedbackResponses = new ArrayList<>();
-
-        feedbackResponses.addAll(load()
-                .filter("feedbackSessionName = ", feedbackSessionName)
-                .filter("courseId =", courseId)
-                .filter("giverSection =", section)
-                .filter("receiverSection =", section)
-                .limit(range + 1).list());
-
-        // also show responses in section with giver but without recipient
-        feedbackResponses.addAll(load()
-                .filter("feedbackSessionName = ", feedbackSessionName)
-                .filter("courseId =", courseId)
-                .filter("giverSection =", section)
-                .filter("receiverSection =", "None")
-                .limit(range + 1).list());
-
-        return feedbackResponses;
     }
 
     private List<FeedbackResponse> getFeedbackResponseEntitiesForSessionFromSectionWithinRange(
